First version of Action.build method
