+++ vizro-core/src/vizro/models/_components/button.py
@@ -6,7 +6,7 @@ from pydantic import Field
 
 from vizro.models import Action, VizroBaseModel
 from vizro.models._action._actions_chain import _action_validator_factory
-from vizro.models._models_utils import _log_call
+from vizro.models._models_utils import _build_component_actions, _log_call
 
 
 class Button(VizroBaseModel):
@@ -29,6 +29,7 @@ class Button(VizroBaseModel):
     def build(self):
         return html.Div(
             [
+                *_build_component_actions(self),
                 dbc.Button(
                     children=self.text,
                     id=self.id,
