+++ vizro-core/src/vizro/models/_components/form/radio_items.py
@@ -6,7 +6,7 @@ from pydantic import Field
 from vizro.models import Action, VizroBaseModel
 from vizro.models._action._actions_chain import _action_validator_factory
 from vizro.models._components.form._form_utils import get_options_and_default
-from vizro.models._models_utils import _log_call
+from vizro.models._models_utils import _build_component_actions, _log_call
 from vizro.models.types import OptionsType, SingleValueType
 
 
@@ -40,6 +40,7 @@ class RadioItems(VizroBaseModel):
 
         return html.Div(
             [
+                *_build_component_actions(self),
                 html.P(self.title, id="radio_items_title") if self.title else None,
                 dcc.RadioItems(
                     id=self.id,
